site/requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
django-rest-framework==0.1.0
PyMySQL==0.8.0
pytz==2018.3
paho-mqtt>=1.3.1
